% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculatePI.R
\name{calculatePI}
\alias{calculatePI}
\title{A function for calculating proliferative index from variance stabalized RNA-seq data in the ProliferativeIndex package}
\usage{
calculatePI(userObject)
}
\arguments{
\item{userObject}{Output from ProliferativeIndex readDataForPI function}
}
\description{
This function allows the user to read in data for subsequent proliferative index calculation and analysis
}
\examples{
calculatePI(userObject, metaPCNA)
# the message "vstData contained x/131 of the PI-associated genes" where x is the number of HGNC gene IDs that matched prints to the console
# the output is calculated proliferative indices for each sample
}

