% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareModeltoPI.R
\name{compareModeltoPI}
\alias{compareModeltoPI}
\title{A function for comparing proliferative indices from variance stabalized RNA-seq data to model principal componenets in the ProliferativeIndex package}
\usage{
compareModeltoPI(userObject, vstPI)
}
\arguments{
\item{userObject}{Output from ProliferativeIndex readDataForPI function (user data)}

\item{vstPI}{Output from ProliferativeIndex calculatePI function}
}
\description{
This function allows the user to calculate a correlation between their model and PI
}
\examples{
compareModeltoPI(userObject, vstPI)
#this function will plot PI vs model PC1 and has a table output that, for the first 10 model PCs, contains the Spearman Rho, Spearman P-values, and PC Proportion of Variance.
}

