% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/comparePI.R
\name{comparePI}
\alias{comparePI}
\title{A function for comparing proliferative indices from variance stabalized RNA-seq data in the ProliferativeIndex package}
\usage{
comparePI(vstPIinput)
}
\arguments{
\item{vstPIinput}{Output from ProliferativeIndex calculatePI function (PI of user vst data)////////////////////////////////////////////////////////////////////////}
}
\description{
This function allows the user to examine a summary and plot of their PIs
}
\examples{
calculatePI(vstPIinput)
#this function summarives the PI values within the user's dataset and plots the PI for each sample
}

