% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exReadDataObj.R
\docType{data}
\name{exReadDataObj}
\alias{exReadDataObj}
\title{TCGA ACC data set output from readDataForPI function}
\format{A list of the vstData (vstTCGA_ACCData_sub data frame with 20501 rows and 10 columns) and the modelIDs (character vector of "AIFM3", "ATP9B", "CTRC", "MCL1", "MGAT4B", "ODF2L", "SNORA65", "TPPP2")}
\usage{
data(exReadDataObj)
}
\description{
This object is the output from readDataForPI taking vstTCGA_ACCData_sub and this model:
c("AIFM3", "ATP9B", "CTRC", "MCL1", "MGAT4B", "ODF2L", "SNORA65", "TPPP2"))
as inputs.
}
\keyword{datasets}
