% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exVSTPI.R
\docType{data}
\name{exVSTPI}
\alias{exVSTPI}
\title{TCGA ACC data set output from calculatePI function}
\format{A numeric vector of the calculated Proliferation Indices for the vstTCGA_ACCData_sub data}
\usage{
data(exVSTPI)
}
\description{
This object is the output from calculatePI from the exReadDataObj as input.
}
\keyword{datasets}
