% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vstTCGA_ACCData_sub.R
\docType{data}
\name{vstTCGA_ACCData_sub}
\alias{vstTCGA_ACCData_sub}
\title{TCGA ACC data set}
\format{A data frame with 20501 rows and 10 variables}
\usage{
data(vstTCGA_ACCData_sub)
}
\description{
A dataset containing data from the The Cancer Genome Atlas (TCGA) Adrenocortical Carcinoma (ACC) dataset. This data was obtained from the TCGA data portal (tcga-data.nci.nih.gov) in June 2015. Level 3 RNASeqV2 raw count data was variance stabalized with the DESeq2 v1.8.2 varianceStabilizingTransformation:
}
\keyword{datasets}

