% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePI.R
\name{calculatePI}
\alias{calculatePI}
\title{A function for calculating proliferative index from variance stabalized RNA-seq data in the ProliferativeIndex package}
\usage{
calculatePI(userObject)
}
\arguments{
\item{userObject}{Output from ProliferativeIndex readDataForPI function}
}
\description{
This function allows the user to read in data for subsequent proliferative index calculation and analysis
}
\examples{
calculatePI(exReadDataObj)
}
