% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDataForPI.R
\name{readDataForPI}
\alias{readDataForPI}
\title{A function for reading data in for use with the ProliferativeIndex package}
\usage{
readDataForPI(vstData, modelIDs)
}
\arguments{
\item{vstData}{Dataframe of user variance stabalized count data (from DESeq2) with samples in columns and genes in rows. Rownames must be genes.}

\item{modelIDs}{Genes in user identified model for comparison to proliferative index}
}
\description{
This function allows the user to read in variance stabalized RNA-seq data and gene model names for subsequent proliferative index calculation and analysis
}
\examples{
readDataForPI(vstTCGA_ACCData_sub, c("AIFM3", "ATP9B", "CTRC", "MCL1", 
"MGAT4B", "ODF2L", "SNORA65", "TPPP2"))
}

